Added support for multiple arch and compilers.

+++ core/enums/test-type-enum.h
@@ -1,8 +1,10 @@
 #pragma once
 
+#include <cstdint>
+
 namespace Enum {
 
-	enum class TestType : unsigned int {
+	enum class TestType : uint8_t {
 		EmptyCall,
 		Fibonacci,
 		NWD,
